requirements.txt
numpy==1.24.3
pandas==1.5.3
pyyaml==6.0
requests==2.29.0
lxml==4.9.2
openpyxl==3.0.10